Added 'neural net params' button and window.
Clicking on the 'Neural Net params' button brings up a window allowing
modification of the free parameters used by the neural net classifier.
The neural net model uses two adjustable parameters:  i) Number of
hidden neurons and ii) Error tolerance.  These are set separately for
the cases of 'rotations' (used when the 'allow rotations' button is
checked and 'no rotations', for when the 'allow rotations' button is
unchecked.